Remove assets, add source to test file + compile
