--- ofrak_core/test_ofrak/components/assets/string_test.out
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e9c6eec22cc30ad1b07f26cf6eecf795a43808021ae8e2116f66b61ab3e1f59f
-size 16720
